Expense Cloning from expense plan to budget.

+++ src/controllers/expense.controller.ts
@@ -1,5 +1,6 @@
 import routeHandler from "express-async-handler";
 import { StatusCodes } from "http-status-codes";
+import omit from "lodash/omit";
 import { PipelineStage, Types } from "mongoose";
 import Expense from "../models/expense.model";
 import ExpensePlan from "../models/expensePlan.model";
@@ -40,8 +41,7 @@ export const createExpense = routeHandler(
 );
 
 /**
- * Update an expense.
- * @description get expenses of a user
+ * @description Updates expense and any linked expense as well.
  * @method PUT /api/expenses
  * @access protected
  */
@@ -58,6 +58,8 @@ export const updateExpense = routeHandler(
       throw new Error("Please provide all required fields.");
     }
 
+    // TODO: there is probably something wrong here.
+    // FIXME: check and fix.
     await Expense.findByIdAndUpdate(_id, { $set: ex });
     const update: IExpense | null = await Expense.findById(_id).populate(
       "categoryId"
@@ -69,36 +71,87 @@ export const updateExpense = routeHandler(
       });
     }
 
+    // Update linked expense, if any.
+    if (update?.linked) {
+      if (!update.plan) {
+        const linkedExpenseInPlan = await Expense.findById(update.linked);
+        await ExpensePlan.findByIdAndUpdate(linkedExpenseInPlan?.plan, {
+          $set: { lastAction: "Expense Updated" },
+        });
+      }
+      await Expense.findByIdAndUpdate(update.linked, {
+        $set: omit(ex, ["_id", "plan", "linked"]),
+      });
+    }
+
     res.json({ message: "Expense updated successfully.", response: update });
   }
 );
 
 /**
- * Delete expense
- * @description Delete a single expense by a user
+ * @description Delete a single expense by a user, also deletes linked expense, if any
  * @method DELETE /api/expenses
  * @access protected
  */
 export const deleteExpense = routeHandler(
   async (req: TypedRequest<{ id: "string" }, {}>, res: TypedResponse) => {
     const expense: IExpense | null = await Expense.findById(req.query.id);
+
+    // update plan state if expense was in a plan or expense was linked to a plan.
     if (expense?.plan) {
       await ExpensePlan.findByIdAndUpdate(expense.plan, {
         $set: { lastAction: "Expense Removed" },
       });
     }
 
-    const deleted = await Expense.deleteOne({
+    // delete linked expense, and update the plan if current/linked expense was in the plan.
+    if (expense?.linked) {
+      if (!expense.plan) {
+        const linkedExpenseInPlan = await Expense.findById(expense.linked);
+        await ExpensePlan.findByIdAndUpdate(linkedExpenseInPlan?.plan, {
+          $set: { lastAction: "Expense Removed" },
+        });
+      }
+      await Expense.deleteOne({
+        user: new Types.ObjectId(req.userId),
+        _id: new Types.ObjectId(expense.linked),
+      });
+    }
+
+    // delete original
+    await Expense.deleteOne({
       user: new Types.ObjectId(req.userId),
       _id: new Types.ObjectId(req.query.id),
     });
 
-    if (deleted.acknowledged && deleted.deletedCount === 1) {
-      res.json({ message: "Expense deleted successfully." });
-    } else if (deleted.deletedCount === 0) {
-      res.status(StatusCodes.NOT_FOUND);
-      throw new Error("The Expense you're trying to delete does not exist.");
-    }
+    res.json({ message: "Expense deleted successfully." });
+  }
+);
+
+/**
+ * @description Copies an expense to the regular budget.
+ * @method PUT /api/expenses/clone
+ * @access protected
+ */
+export const cloneExpense = routeHandler(
+  async (req: TypedRequest<{}, { _id: string }>, res: TypedResponse) => {
+    const existing = await Expense.findById(req.body._id);
+    const newExpense = await Expense.create({
+      amount: existing?.amount,
+      categoryId: existing?.categoryId,
+      date: existing?.date,
+      description: existing?.description,
+      plan: null,
+      reverted: existing?.reverted,
+      title: existing?.title,
+      user: existing?.user,
+      linked: new Types.ObjectId(existing?._id),
+    });
+
+    existing?.set("linked", new Types.ObjectId(newExpense._id));
+    existing?.save();
+
+    res.json({ message: "Expense copied to budget successfully!" });
   }
 );
